This is the revised version after the review

+++ NunitDemo/src/Tests/BookTest/DeleteBookTest.cs
@@ -47,8 +47,11 @@ public class DeleteBookTest : BaseTest
     [TearDown]
     public async Task Teardown()
     {
-        var body = new DeleteAndReplaceBookModel{isbn = _isbn, userId = UserId};
-        await BookService.DeleteBookFromCollection(body: body, token: Token);
+        if (!string.IsNullOrEmpty(UserId) && !string.IsNullOrEmpty(_isbn))
+        {
+            var body = new DeleteAndReplaceBookModel{isbn = _isbn, userId = UserId};
+            await BookService.DeleteBookFromCollection(body: body, token: Token);        
+        }
     }
 
 }
